# _refs/openapi-soundcloud-com.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SoundCloud Public API Specification",
    "description": "Swagger json can be found [here](https://developers.soundcloud.com/docs/api/explorer/api.json)\n\n\u003Ca href=\"#\" id=\"cookie-manager\"\u003ECookie manager\u003C/a\u003E\n",
    "contact": {
      "name": "API issue tracker",
      "url": "https://github.com/soundcloud/api"
    },
    "version": "1.0.0",
    "termsOfService": "https://developers.soundcloud.com/docs/api/terms-of-use"
  },
  "servers": [
    {
      "url": "https://api.soundcloud.com"
    }
  ],
  "security": [
    {
      "AuthHeader": []
    },
    {
      "oAuth2_1": []
    }
  ],
...
  "paths": {
    "/me": {
      "get": {
        "summary": "Returns the authenticated user’s information.",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Me"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/me/feed": {
      "get": {
        "summary": "Returns the authenticated user's feed.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Activities"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/me/feed/tracks": {
      "get": {
        "summary": "Returns the authenticated user's recent track related feed.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Activities"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/me/likes/tracks": {
      "get": {
        "summary": "Returns a list of favorited or liked tracks of the authenticated user.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/access_explicit"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Tracks"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/me/likes/playlists": {
      "get": {
        "summary": "Returns a list of favorited or liked playlists of the authenticated user.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Playlists"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/me/followings": {
      "get": {
        "summary": "Returns a list of users who are followed by the authenticated user.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Users"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/me/followings/tracks": {
      "get": {
        "summary": "Returns a list of recent tracks from users followed by the authenticated user.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TracksList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/me/followings/{user_urn}": {
      "get": {
        "summary": "Returns a user who is followed by the authenticated user. (use /users/{user_urn} instead, to fetch the user details)",
        "deprecated": true,
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/User"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Follows a user.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn_to_follow"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Successful"
          },
          "201": {
            "$ref": "#/components/responses/User"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Deletes a user who is followed by the authenticated user.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn_to_follow"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json; charset=utf-8": {
                "examples": {
                  "200": {
                    "$ref": "#/components/examples/OK"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          }
        }
      }
    },
    "/me/followers": {
      "get": {
        "summary": "Returns a list of users who are following the authenticated user.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Users"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/me/playlists": {
      "get": {
        "summary": "Returns user’s playlists (sets).",
        "description": "Returns playlist info, playlist tracks and tracks owner info.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/show_tracks"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Playlists"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/me/tracks": {
      "get": {
        "summary": "Returns a list of user's tracks.",
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Tracks"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/tracks": {
      "get": {
        "summary": "Performs a track search based on a query",
        "tags": [
          "search"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/ids"
          },
          {
            "$ref": "#/components/parameters/urns"
          },
          {
            "$ref": "#/components/parameters/genres"
          },
          {
            "$ref": "#/components/parameters/tags"
          },
          {
            "$ref": "#/components/parameters/bpm"
          },
          {
            "$ref": "#/components/parameters/duration"
          },
          {
            "$ref": "#/components/parameters/created_at"
          },
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Tracks"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Uploads a new track.",
        "tags": [
          "tracks"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TrackDataRequest"
                  },
                  {
                    "required": [
                      "track[title]",
                      "track[asset_data]"
                    ]
                  }
                ]
              },
              "examples": {
                "default": {
                  "$ref": "#/components/examples/TrackDataRequest"
                }
              }
            },
            "multipart/x-www-form-urlencoded": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TrackDataRequest"
                  },
                  {
                    "required": [
                      "track[title]",
                      "track[asset_data]"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Track"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          }
        }
      }
    },
    "/playlists": {
      "get": {
        "summary": "Performs a playlist search based on a query",
        "tags": [
          "search"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/show_tracks"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Playlists"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Creates a playlist.",
        "tags": [
          "playlists"
        ],
        "requestBody": {
          "description": "Create Playlist request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdatePlaylistRequest"
              },
              "examples": {
                "default": {
                  "$ref": "#/components/examples/CreateUpdatePlaylistRequest"
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateUpdatePlaylistFormRequest"
                  },
                  {
                    "required": [
                      "playlist[title]"
                    ]
                  }
                ]
              },
              "examples": {
                "default": {
                  "$ref": "#/components/examples/CreateUpdatePlaylistFormRequest"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Playlist"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "bpm": {
        "in": "query",
        "name": "bpm",
        "description": "Return tracks with a specified bpm[from], bpm[to]",
        "required": false,
        "style": "deepObject",
        "schema": {
          "type": "object",
          "properties": {
            "from": {
              "type": "integer",
              "description": "Return tracks with at least this bpm value",
              "example": 123
            },
            "to": {
              "type": "integer",
              "description": "Return tracks with at most this bpm value",
              "example": 456
            }
          }
        }
      },
...
      "created_at": {
        "in": "query",
        "name": "created_at",
        "description": "(yyyy-mm-dd hh:mm:ss) return tracks created within the specified dates",
        "required": false,
        "style": "deepObject",
        "schema": {
          "type": "object",
          "properties": {
            "from": {
              "type": "string",
              "description": "(yyyy-mm-dd hh:mm:ss) return tracks created at this date or later",
              "example": "2020-12-24 00:00:00"
            },
            "to": {
              "type": "string",
              "description": "(yyyy-mm-dd hh:mm:ss) return tracks created at this date or earlier",
              "example": "2020-12-26 00:00:00"
            }
          }
        }
      },
      "duration": {
        "in": "query",
        "name": "duration",
        "description": "Return tracks within a specified duration range",
        "required": false,
        "style": "deepObject",
        "schema": {
          "type": "object",
          "properties": {
            "from": {
              "type": "integer",
              "description": "Return tracks with at least this duration value",
              "example": 123456
            },
            "to": {
              "type": "integer",
              "description": "Return tracks with at most this duration value",
              "example": 456789
            }
          }
        }
      },
      "genres": {
        "in": "query",
        "name": "genres",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A comma separated list of genres",
        "example": "Pop,House"
      },
      "ids": {
        "in": "query",
        "name": "ids",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A comma separated list of track ids to filter on",
        "example": "1,2,3"
      },
      "urns": {
        "in": "query",
        "name": "urns",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A comma separated list of track urns to filter on",
        "example": "soundcloud:tracks:1,soundcloud:tracks:2,soundcloud:tracks:3"
      },
      "limit": {
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        },
        "description": "Number of results to return in the collection.",
        "example": 2
      },
      "offset": {
        "deprecated": true,
        "in": "query",
        "name": "offset",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 0
        },
        "description": "Offset of first result. Deprecated, use `linked_partitioning` instead.",
        "example": 0
      },
      "q": {
        "in": "query",
        "name": "q",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "search",
        "example": "hello"
      },
...
      "tags": {
        "in": "query",
        "name": "tags",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A comma separated list of tags",
        "example": "test"
      },
...
      "linked_partitioning": {
        "in": "query",
        "name": "linked_partitioning",
        "required": false,
        "schema": {
          "type": "boolean"
        },
        "description": "Returns paginated collection of items (recommended, returning a list without pagination is deprecated and should not be used)",
        "example": true
      },
...
      "access": {
        "in": "query",
        "name": "access",
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "playable",
              "preview",
              "blocked"
            ]
          },
          "default": "playable,preview"
        },
        "description": "Filters content by level of access the user (logged in or anonymous) has to the track. The result list will include only tracks with the specified access. Include all options if you'd like to see all possible tracks. See `Track#access` schema for more details.\n",
        "style": "form",
        "explode": false
      },
      "access_explicit": {
        "in": "query",
        "name": "access",
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "playable",
              "preview",
              "blocked"
            ]
          },
          "default": "playable,preview,blocked"
        },
        "description": "Filters content by level of access the user (logged in or anonymous) has to the track. The result list will include only tracks with the specified access. Include all options if you'd like to see all possible tracks. See `Track#access` schema for more details.\n",
        "style": "form",
        "explode": false
      },
      "show_tracks": {
        "in": "query",
        "name": "show_tracks",
        "required": false,
        "schema": {
          "type": "boolean"
        },
        "description": "A boolean flag to request a playlist with or without tracks. Default is `true`.",
        "example": true
      },
...
      "user_urn": {
        "in": "path",
        "name": "user_urn",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "SoundCloud User id",
        "example": "soundcloud:users:948745750"
      },
      "user_urn_to_follow": {
        "in": "path",
        "name": "user_urn",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "SoundCloud User id",
        "example": "soundcloud:users:743372812"
      }
    },
    "securitySchemes": {
      "AuthHeader": {
        "in": "header",
        "type": "apiKey",
        "name": "Authorization",
        "description": "The string \"OAuth\" followed by the user’s token: OAuth ACCESS_TOKEN. Obtained in oAuth2.1 flow"
      },
      "oAuth2_1": {
        "type": "oauth2",
        "description": "This API uses OAuth2.1 with Authorization Code flow, Client Credential flow. When registering your app, for it to work with Swagger UI specify redirect url as https://developers.soundcloud.com/docs/api/explorer/oauth2-redirect.html. [More info](https://developers.soundcloud.com/docs/api/guide#authentication)",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://secure.soundcloud.com/authorize",
            "tokenUrl": "https://secure.soundcloud.com/oauth/token",
            "refreshUrl": "https://secure.soundcloud.com/oauth/token",
            "scopes": {

            }
          },
          "clientCredentials": {
            "tokenUrl": "https://secure.soundcloud.com/oauth/token",
            "scopes": {

            }
          }
        }
      }
    },
    "schemas": {
...
      "TrackDataRequest": {
        "type": "object",
        "properties": {
          "track[title]": {
            "type": "string"
          },
          "track[asset_data]": {
            "title": "Only for uploading",
            "type": "string",
            "format": "binary"
          },
          "track[permalink]": {
            "type": "string"
          },
          "track[sharing]": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ],
            "default": "public"
          },
          "track[embeddable_by]": {
            "type": "string",
            "enum": [
              "all",
              "me",
              "none"
            ],
            "description": "who can embed this track \"all\", \"me\", or \"none\""
          },
          "track[purchase_url]": {
            "type": "string"
          },
          "track[description]": {
            "type": "string"
          },
          "track[genre]": {
            "type": "string"
          },
          "track[tag_list]": {
            "type": "string",
            "description": "The tag_list property contains a list of tags separated by spaces. Multiword tags are quoted in double quotes. We also support machine tags that follow the pattern NAMESPACE:KEY=VALUE. For example: geo:lat=43.555 camel:size=medium “machine:tag=with space” Machine tags are not revealed to the user on the track pages."
          },
          "track[label_name]": {
            "type": "string"
          },
          "track[release]": {
            "type": "string"
          },
          "track[release_date]": {
            "title": "write only!",
            "type": "string",
            "description": "string, formatted as yyyy-mm-dd, representing release date"
          },
          "track[streamable]": {
            "type": "boolean",
            "default": true
          },
          "track[downloadable]": {
            "type": "boolean",
            "default": true
          },
          "track[license]": {
            "type": "string",
            "enum": [
              "no-rights-reserved",
              "all-rights-reserved",
              "cc-by",
              "cc-by-nc",
              "cc-by-nd",
              "cc-by-sa",
              "cc-by-nc-nd",
              "cc-by-nc-sa"
            ],
            "description": "Possible values: no-rights-reserved, all-rights-reserved, cc-by, cc-by-nc, cc-by-nd, cc-by-sa, cc-by-nc-nd, cc-by-nc-sa"
          },
          "track[commentable]": {
            "type": "boolean",
            "default": true
          },
          "track[isrc]": {
            "type": "string"
          },
          "track[artwork_data]": {
            "title": "Only for uploading, for PRO users",
            "type": "string",
            "format": "binary"
          }
        }
      },
...
    },
    "responses": {
...
      "Successful": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/Successful"
              }
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json; charset=utf-8": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/BadRequest"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json; charset=utf-8": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/Unauthorized"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/NotFound"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Unprocessable Entity",
        "content": {
          "application/json": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/UnprocessableEntity"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "User": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/User"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/User"
              }
            }
          }
        }
      },
      "Me": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/Me"
            }
          }
        }
      },
      "Users": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/Users"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Users"
              }
            }
          }
        }
      },
      "Track": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/Track"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Track"
              }
            }
          }
        }
      },
      "Tracks": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Tracks"
                },
                {
                  "$ref": "#/components/schemas/TracksList"
                }
              ]
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Tracks"
              },
              "tracksList": {
                "$ref": "#/components/examples/TracksList"
              }
            }
          }
        }
      },
      "TracksList": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/TracksList"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/TracksList"
              }
            }
          }
        }
      },
...
      "Playlists": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Playlists"
                },
                {
                  "$ref": "#/components/schemas/PlaylistsArray"
                }
              ]
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Playlists"
              },
              "playlistsArray": {
                "$ref": "#/components/examples/PlaylistsArray"
              }
            }
          }
        }
      },
      "Activities": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/Activities"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Activities"
              }
            }
          }
        }
      },
...
    },
    "examples": {
...
      "CreateUpdatePlaylistRequest": {
        "value": {
          "playlist": {
            "title": "Test Auto-created Playlist",
            "description": "Playlist for test purposes",
            "sharing": "private",
            "tracks": [
              {
                "urn": "soundcloud:tracks:219787221"
              },
              {
                "urn": "soundcloud:tracks:783019264"
              },
              {
                "urn": "soundcloud:tracks:870073492"
              }
            ]
          }
        }
      },
      "CreateUpdatePlaylistFormRequest": {
        "value": {
          "playlist[title]": "Test drum sample",
          "playlist[sharing]": "private",
          "playlist[description]": "Public-api test sample upload"
        }
      },
      "TrackDataRequest": {
        "value": {
          "track[title]": "Test drum sample",
          "track[asset_data]": "\u003Cbinary string\u003E",
          "track[permalink]": "test-drum-sample",
          "track[sharing]": "private",
          "track[embeddable_by]": "me",
          "track[purchase_url]": null,
          "track[description]": "Public-api test sample upload",
          "track[genre]": "Pop",
          "track[tag_list]": null,
          "track[label_name]": null,
          "track[release]": null,
          "track[release_date]": "2021-02-18T00:00:00.000Z",
          "track[streamable]": true,
          "track[downloadable]": false,
          "track[license]": "no-rights-reserved",
          "track[commentable]": true,
          "track[isrc]": null,
          "track[artwork_data]": null
        }
      },
...
      "Successful": {
        "value": {
          "status": "200 - Successful"
        }
      },
...
      "Unauthorized": {
        "value": {
          "code": 401,
          "message": "Detailed message of errors, when available",
          "link": "https://developers.soundcloud.com/docs/api/explorer/open-api"
        }
      },
...
      "OK": {
        "value": {
          "status": "Status(200) - OK"
        }
      },
...
    }
  }
}